Added PAPI support. When PAPI is enabled, all timers also collect PAPI_TOT_INS, PAPI_TOT_CYC and PAPI_L2_TCM. Nothing is done with them currently. At the end of execution, all PAPI counters are summed over all worker threads, and they are output if APEX_SCREEN_OUTPUT is set to 1.

+++ apex/thread_instance.hpp
@@ -44,7 +44,7 @@ private:
 public:
   profiler * current_timer;
   static thread_instance& instance(void);
-  static int get_id(void) { return instance()._id; }
+  static long unsigned int get_id(void) { return instance()._id; }
   static std::string get_name(void);
   static void set_name(std::string name);
   static void set_worker(bool is_worker);
